Moved contents of CoreDSLLiftCFToSCF into CoreDSLSwitchToIf and used it in CoreDSLToPy (because of this, deleted conversion code for scf.index_switch)
