+++ lib/Conversion/CMakeLists.txt
@@ -2,4 +2,3 @@ add_subdirectory(AnalyzeISAX)
 add_subdirectory(CoreDSLToPy)
 add_subdirectory(CoreDSLSwitchToIf)
 add_subdirectory(MergeISAX)
-add_subdirectory(CoreDSLLiftCFToSCF)
