+++ lib/Conversion/CoreDSLSwitchToIf/CoreDSLSwitchToIf.cpp
@@ -5,6 +5,7 @@
 #include "circt/Dialect/HW/HWDialect.h"
 #include "circt/Dialect/HWArith/HWArithDialect.h"
 #include "circt/Dialect/HWArith/HWArithOps.h"
+#include "mlir/Conversion/ControlFlowToSCF/ControlFlowToSCF.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Transforms/Passes.h"
 
@@ -82,13 +83,75 @@ namespace {
 struct CoreDSLSwitchToIf
     : public mlir::shortnail::impl::CoreDSLSwitchToIfBase<CoreDSLSwitchToIf> {
   using CoreDSLSwitchToIfBase::CoreDSLSwitchToIfBase;
+  // This needs to be a class rather than a lambda because we cannot have
+  // templated lambdas and this needs to be explicitly instantiated with
+  // multiple op types for the walk method
+  template <typename OpType>
+  struct RegionToSCFConverter {
+    // So we can use Pass::getChildAnalysis
+    friend struct CoreDSLSwitchToIf;
+    CoreDSLSwitchToIf &pass;
+    bool &changed;
+
+    WalkResult operator()(OpType op) {
+      ControlFlowToSCFTransformation transformation;
+      DominanceInfo &dominanceInfo = pass.getChildAnalysis<DominanceInfo>(op);
+      FailureOr<bool> res =
+          transformCFGToSCF(op.getRegion(), transformation, dominanceInfo);
+      if (failed(res)) {
+        op->emitError("Failed to convert all cf ops to scf");
+        return WalkResult::interrupt();
+      }
+      changed |= *res;
+      return WalkResult::skip();
+    }
+  };
 
   void runOnOperation() override {
     auto isaxOp = getOperation();
+    // Visit all operations that can have multiple child blocks. As SCF ops
+    // other than scf::ExecuteRegionOp cannot have multiple child blocks, they
+    // don't need to be visited here
+    bool changed = false;
+    static constexpr WalkOrder ORDER = WalkOrder::PostOrder;
+    // NOTE: Treenail currently always generates an scf.execute_region operation
+    // around every cf.switch. Because of this, it is technically only necessary
+    // to run this on scf::ExecuteRegionOp and func::FuncOp
+    auto res = isaxOp->walk<ORDER>(
+        RegionToSCFConverter<coredsl::InstructionOp>{*this, changed});
+    if (res == WalkResult::interrupt()) {
+      return signalPassFailure();
+    }
+    res = isaxOp->walk<ORDER>(
+        RegionToSCFConverter<coredsl::AlwaysOp>{*this, changed});
+    if (res == WalkResult::interrupt()) {
+      return signalPassFailure();
+    }
+    res = isaxOp->walk<ORDER>(
+        RegionToSCFConverter<coredsl::SpawnOp>{*this, changed});
+    if (res == WalkResult::interrupt()) {
+      return signalPassFailure();
+    }
+    res =
+        isaxOp->walk<ORDER>(RegionToSCFConverter<func::FuncOp>{*this, changed});
+    if (res == WalkResult::interrupt()) {
+      return signalPassFailure();
+    }
+    res = isaxOp->walk<ORDER>(
+        RegionToSCFConverter<scf::ExecuteRegionOp>{*this, changed});
+    if (res == WalkResult::interrupt()) {
+      return signalPassFailure();
+    }
+    if (!changed) {
+      markAllAnalysesPreserved();
+      // If nothing changed, there is no need to convert scf.index_switch to
+      // scf.if, as we expect there to not be any index switch ops other than
+      // the one generated by converting cf.switch ops
+      return;
+    }
     MLIRContext &ctx = getContext();
     // First convert the cf operations
     OpPassManager cfToSCFPM{isaxOp.getOperationName()};
-    cfToSCFPM.addPass(createCoreDSLLiftCFToSCF());
     if (failed(runPipeline(cfToSCFPM, isaxOp))) {
       return signalPassFailure();
     }
