+++ lib/Conversion/CoreDSLToPy/CMakeLists.txt
@@ -14,5 +14,5 @@ add_shortnail_library(ShortnailCoreDSLToPy
     CIRCTSupport
 
     ShortnailCoreDSL
-    ShortnailCoreDSLLiftCFToSCF
+    ShortnailCoreDSLSwitchToIf
 )
