+++ lib/Conversion/CoreDSLToPy/CoreDSLToPy.cpp
@@ -607,7 +607,6 @@ static WalkResult emitArithOp(mlir::raw_indented_ostream &os, Operation *op) {
 static WalkResult emitSCFOp(mlir::raw_indented_ostream &os, Operation *op) {
   static unsigned uniqueIfNumber = 0;
   static unsigned uniqueForNumber = 0;
-  static unsigned uniqueSwitchNumber = 0;
 
   assert(isa<scf::SCFDialect>(op->getDialect()));
 
@@ -766,79 +765,6 @@ static WalkResult emitSCFOp(mlir::raw_indented_ostream &os, Operation *op) {
 
         return WalkResult::skip();
       })
-      // NOTE: Only generated by ControlFlowToSCF
-      .Case<scf::IndexSwitchOp>([&](scf::IndexSwitchOp switchOp) {
-        const unsigned switchId = uniqueSwitchNumber++;
-        for (unsigned i = 0, end = switchOp.getNumCases(); i < end; ++i) {
-          Block &caseBlock = switchOp.getCaseBlock(i);
-          // Emit each case as a helper function. This way each
-          // switch has the form
-          // 'ret1, ret2, ..., retn = helper_function', thereby
-          // making handling of the results easier and allowing
-          // us to just emit a return for every yield op
-          os << "def helper_function_coredsl_switch_" << switchId << "_case_"
-             << i << "():\n";
-          os.indent();
-          auto res = caseBlock.template walk<WalkOrder::PreOrder>(
-              [&os](Operation *op) { return emitOp(os, op); });
-          if (res.wasInterrupted())
-            return res;
-          os.unindent();
-        }
-        os << "def helper_function_coresl_switch_" << switchId
-           << "_default_case():\n";
-        os.indent();
-        auto res = switchOp.getDefaultBlock().walk<WalkOrder::PreOrder>(
-            [&os](Operation *op) { return emitOp(os, op); });
-        if (res.wasInterrupted())
-          return res;
-        os.unindent();
-        bool first = true;
-        for (const auto &[idx, attr] : llvm::enumerate(switchOp.getCases())) {
-          const int64_t caseVal = attr;
-          if (first)
-            os << "if ";
-          else
-            os << "elif ";
-          first = false;
-          valToPy(os, switchOp.getArg());
-          os << "." << hwarith::ICmpPredicate::eq << "(";
-          os << caseVal << "):\n";
-          os.indent();
-          if (switchOp.getNumResults() > 0) {
-            bool first = true;
-            for (auto res : switchOp.getResults()) {
-              if (!first) {
-                os << ", ";
-              }
-              first = false;
-              valToPy(os, res);
-            }
-            os << " = ";
-          }
-          os << "helper_function_coredsl_switch_" << switchId << "_case_" << idx
-             << "()\n";
-          os.unindent();
-        }
-        os << "else:\n";
-        os.indent();
-        if (switchOp.getNumResults() > 0) {
-          bool first = true;
-          for (auto res : switchOp.getResults()) {
-            if (!first) {
-              os << ", ";
-            }
-            first = false;
-            valToPy(os, res);
-          }
-          os << " = ";
-        }
-        os << "helper_function_coredsl_switch_" << switchId
-           << "_default_case()\n";
-        os.unindent();
-        // Skip the child scopes, as we already visited them
-        return WalkResult::skip();
-      })
       .Default([&](auto _) {
         op->emitError()
             << "CoreDSLToPy::emitSCFOp lacks emission code for this operation!";
@@ -927,7 +853,7 @@ struct CoreDSLToPy
 
     OpPassManager inlinePM(isaxOp.getOperationName());
     // Before inliing, convert all cf operations to scf
-    inlinePM.addPass(createCoreDSLLiftCFToSCF());
+    inlinePM.addPass(createCoreDSLSwitchToIf());
     // Inline all functions into the InstructionOps
     inlinePM.addPass(createInlinerPass());
     if (failed(runPipeline(inlinePM, isaxOp)))
